+++ UltimateCastbars/CHANGELOGS.lua
@@ -1,6 +1,14 @@
 local _, UCB = ...
 
 UCB.CHANGELOG_TEXT = [=[
+## Version 0.9.4 - [26-02-2026]
+
+### Fixed
+- Channel casts that were secrets caused issues with cancelled overlays
+- Alpha for unkickable and uninterruptible casts at the same time were causing unwanted behaviour
+
+---
+
 ## Version 0.9.3 - [26-02-2026]
 
 ### Fixed