+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Fixed
 - Channel casts that were secrets caused issues with cancelled overlays
+- Alpha for unkickable and uninterruptible casts at the same time were causing unwanted behaviour
 
 ---