+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased
 
+---
+
+
+## Version 0.9.4 - [26-02-2026]
+
 ### Fixed
 - Channel casts that were secrets caused issues with cancelled overlays
 - Alpha for unkickable and uninterruptible casts at the same time were causing unwanted behaviour